requirements.txt
Authlib==1.0.1
bcrypt==3.2.2
boto3==1.24.27
botocore==1.27.27
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
cryptography==37.0.4
Flask==2.1.2
Flask-Login==0.6.1
google-api-core==2.8.2
google-api-python-client==2.52.0
google-auth==2.9.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.3
httplib2==0.20.4
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.1
oauthlib==3.2.0
packaging==21.3
paramiko==2.11.0
protobuf==4.21.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyMySQL==1.0.2
PyNaCl==1.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
rauth==0.7.3
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.8
s3transfer==0.6.0
six==1.16.0
sshtunnel==0.4.0
uritemplate==4.1.1
urllib3==1.26.10
vendor==0.1
Werkzeug==2.1.2